requirements.txt
streamlit
streamlit-aggrid==0.2.2-2
openai
tqdm
pandas
python-dotenv